pyproject.toml
[tool.poetry]
name = "taxes"
version = "0.1.0"
description = ""
authors = ["Nick Webster <nick@nick.geek.nz>"]

[tool.poetry.dependencies]
python = "^3.9"
py-money = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
